share interface and relocate osaka cm data
