+++ src/data/exhibition/osaka/creators-market/booths.ts
@@ -1,25 +1,4 @@
-interface MemberLink {
-  label: string;
-  url: string;
-}
-
-interface Member {
-  name: string;
-  koName?: string;
-  links?: MemberLink[];
-}
-
-export interface Booth {
-  id: string;
-  row: string;
-  col: number;
-  span?: number;
-  name: string;
-  koPNames?: string;
-  dates: string[];
-  members: Member[];
-  hidden?: boolean;
-}
+import type { Booth } from '@/types/booth';
 
 const boothsA: Booth[] = [
   {
