+++ src/data/exhibition/tokyo/creators-market/constants.ts
@@ -1,0 +1,5 @@
+export const DAYS = [] as const;
+export const jacketSrc = (id: string) => `/images/tokyo/creators-market/cc_${id}.jpg`;
+export const displayBoothId = (id: string) => id;
+export const rowColors: Record<string, string> = {};
+export const dayClass: Record<string, string> = {};
